requirements.txt
dj-database-url==0.4.2
Django==1.11.4
djangorestframework==3.6.3
gunicorn==19.7.1
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
whitenoise==3.3.0
